Refactor : 회원가입 검증, 로그인 관련 테스트 메서드 이동
